+++ CMakeLists.txt
@@ -164,6 +164,8 @@ add_library(rdf4cpp
         private/rdf4cpp/parser/XMLStates/XMLParserXMLLiteralState.cpp
         private/rdf4cpp/parser/JsonLdParser.cpp
         private/rdf4cpp/parser/JsonLdParserTypes.cpp
+        private/rdf4cpp/parser/JsonLdContextParser.cpp
+        private/rdf4cpp/parser/JsonLdExpandParser.cpp
         private/rdf4cpp/regex/RegexImpl.cpp
         private/rdf4cpp/regex/RegexReplacerImpl.cpp
         ${serd_source_files}
